// src/customers/booking/dto/booking-status.dto.ts
export enum BookingStatus {
  CONFIRMED = 'CONFIRMED',
  SEARCHING_DRIVER = 'SEARCHING_DRIVER',
  DRIVER_REQUESTED = 'DRIVER_REQUESTED',
  DRIVER_NOTIFIED = 'DRIVER_NOTIFIED',
  DRIVER_ACCEPTED = 'DRIVER_ACCEPTED',
  DRIVER_ARRIVING = 'DRIVER_ARRIVING',
  DRIVER_ASSIGNED = 'DRIVER_ASSIGNED',
  TRIP_STARTED = 'TRIP_STARTED',
  TRIP_COMPLETED = 'TRIP_COMPLETED',
  PAYMENT_PENDING = 'PAYMENT_PENDING',
  PAID = 'PAID',
  CANCELLED = 'CANCELLED',
  NO_DRIVER_FOUND = 'NO_DRIVER_FOUND',
  COMPLETED = 'COMPLETED',
}